requirements.txt
Flask==2.2.3 
gunicorn==20.1.0
Werkzeug==2.2.2
requests==2.28.2
tensorflow==2.16.2
mlflow<3,>=2.2
pillow==10.4.0